+++ metatensor-core/include/metatensor.hpp
@@ -962,9 +962,11 @@ namespace details {
 namespace io {
     /// Save a `TensorMap` to the file at `path`.
     ///
-    /// If the file exists, it will be overwritten.
+    /// If the file exists, it will be overwritten. The recomended file
+    /// extension when saving data is `.mts`, to prevent confusion with generic
+    /// `.npz` files.
     ///
-    /// `TensorMap` are serialized using numpy's `.npz` format, i.e. a ZIP file
+    /// `TensorMap` are serialized using numpy's NPZ format, i.e. a ZIP file
     /// without compression (storage method is `STORED`), where each file is
     /// stored as a `.npy` array. See the C API documentation for more
     /// information on the format.
@@ -984,7 +986,9 @@ namespace io {
 
     /// Save a `TensorBlock` to the file at `path`.
     ///
-    /// If the file exists, it will be overwritten.
+    /// If the file exists, it will be overwritten. The recomended file
+    /// extension when saving data is `.mts`, to prevent confusion with generic
+    /// `.npz` files.
     void save(const std::string& path, const TensorBlock& block);
 
     /// Save a `TensorBlock` to an in-memory buffer.
@@ -1001,7 +1005,9 @@ namespace io {
 
     /// Save `Labels` to the file at `path`.
     ///
-    /// If the file exists, it will be overwritten.
+    /// If the file exists, it will be overwritten. The recomended file
+    /// extension when saving data is `.mts`, to prevent confusion with generic
+    /// `.npz` files.
     void save(const std::string& path, const Labels& labels);
 
     /// Save `Labels` to an in-memory buffer.
@@ -1029,7 +1035,7 @@ namespace io {
      *
      * \endverbatim
      *
-     * `TensorMap` are serialized using numpy's `.npz` format, i.e. a ZIP file
+     * `TensorMap` are serialized using numpy's NPZ format, i.e. a ZIP file
      * without compression (storage method is `STORED`), where each file is
      * stored as a `.npy` array. See the C API documentation for more
      * information on the format.