+++ python/metatensor_operations/tests/equal_metadata.py
@@ -12,12 +12,12 @@ DATA_ROOT = os.path.join(os.path.dirname(__file__), "data")
 
 @pytest.fixture
 def test_tensor_map_1() -> TensorMap:
-    return metatensor.load(os.path.join(DATA_ROOT, "qm7-power-spectrum.npz"))
+    return metatensor.load(os.path.join(DATA_ROOT, "qm7-power-spectrum.mts"))
 
 
 @pytest.fixture
 def test_tensor_map_2() -> TensorMap:
-    return metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
+    return metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.mts"))
 
 
 @pytest.fixture