+++ python/metatensor_operations/tests/split.py
@@ -12,7 +12,7 @@ DATA_ROOT = os.path.join(os.path.dirname(__file__), "data")
 
 def test_split_block_samples():
     # use a TensorMap with multiple components
-    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
+    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.mts"))
     block = tensor.block(o3_lambda=2, center_type=6, neighbor_type=6)
 
     selections = [
@@ -60,7 +60,7 @@ def test_split_block_samples():
 
 def test_split_block_samples_not_everything():
     # use a TensorMap with multiple components
-    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
+    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.mts"))
     block = tensor.block(o3_lambda=2, center_type=6, neighbor_type=6)
 
     # using `selections` with some samples not present in the initial block,
@@ -89,7 +89,7 @@ def test_split_block_samples_not_everything():
 
 
 def test_split_samples():
-    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
+    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.mts"))
 
     selections = [
         Labels(names=["system"], values=np.array([[0], [6], [7]])),
@@ -124,7 +124,7 @@ def test_split_samples():
 
 def test_split_block_properties():
     # TensorMap with multiple properties
-    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-power-spectrum.npz"))
+    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-power-spectrum.mts"))
     block = tensor.block(center_type=8, neighbor_1_type=6, neighbor_2_type=8)
 
     selections = [
@@ -166,7 +166,7 @@ def test_split_block_properties():
 
 def test_split_properties():
     # TensorMap with multiple properties
-    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-power-spectrum.npz"))
+    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-power-spectrum.mts"))
 
     selections = [
         Labels(names=["l", "n_2"], values=np.array([[0, 0], [1, 3], [3, 1]])),
@@ -192,7 +192,7 @@ def test_split_properties():
 
 
 def test_split_errors():
-    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
+    tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.mts"))
     block = tensor.block(4)
     selections = [
         Labels(names=["system"], values=np.array([[0], [6], [7]])),