+++ python/metatensor_torch/tests/operations/add.py
@@ -19,7 +19,7 @@ def tensor_path():
         "metatensor_operations",
         "tests",
         "data",
-        "qm7-power-spectrum.npz",
+        "qm7-power-spectrum.mts",
     )